requirements.txt

numpy~=1.21.2
parameterized~=0.8.1

python-constraint~=1.4.0
pandas~=1.3.3